backend/requirements.txt
flask
flask_sqlalchemy
flask_jwt_extended
psycopg2-binary
flask_cors